data/dataset.py
import os
from PIL import Image
import torch
from torch.utils.data import Dataset
from torchvision import transforms

class BloodSmearDataset(Dataset):
    """
    Custom Dataset for Microscopic Blood Smear Images (Leukemia).
    Applies noise reduction, color normalization implicitly via ImageNet stats or custom stats,
    and augmentations for training.
    """
    def __init__(self, data_dir, split="train", img_size=224):
        self.data_dir = data_dir
        self.split = split
        self.img_size = img_size
        
        # Example directory structure: data_dir / split / class_name / image.jpg
        self.image_paths = []
        self.labels = []
        self.class_to_idx = {"healthy": 0, "leukemia": 1} # Customize based on ALL-IDB/C-NMC
        
        # Dummy initialization for structural scaffolding since datasets aren't downloaded
        self._mock_data()
        
        self.transform = self._get_transforms()
        
    def _mock_data(self):
        # Mocks dataset if actual data isn't present
        print(f"Loading {self.split} dataset...")
        # In a real scenario, we'd walk the directory structure:
        # for class_name in os.listdir(os.path.join(self.data_dir, self.split)):
        #     ...
        pass

    def _get_transforms(self):
        # Color normalization using ImageNet means/stds usually sufficient for pretrained ViTs
        normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406],
                                         std=[0.229, 0.224, 0.225])
        
        if self.split == "train":
            return transforms.Compose([
                transforms.Resize((self.img_size, self.img_size)),
                transforms.RandomHorizontalFlip(),
                transforms.RandomVerticalFlip(),
                transforms.RandomRotation(15),
                transforms.ColorJitter(brightness=0.1, contrast=0.1),
                transforms.ToTensor(),
                normalize
            ])
        else:
            return transforms.Compose([
                transforms.Resize((self.img_size, self.img_size)),
                transforms.ToTensor(),
                normalize
            ])

    def __len__(self):
        # Return mocked length
        return 100 if self.split == "train" else 20

    def __getitem__(self, idx):
        # Dummy data generation for testing scaffolding
        # In reality, load PIL Image and apply self.transform
        img = torch.randn(3, self.img_size, self.img_size) 
        label = torch.randint(0, 2, (1,)).item()
        return img, label
